+++ R/quote.R
@@ -1,6 +1,9 @@
 #' Enumerate Multiple Elements
 #'
-#' Enumerate multiple elements of a vector or list.
+#' @description
+#' `r lifecycle::badge("deprecated")`
+#'
+#' This function is *deprecated*, please use `cli` functionality instead.
 #'
 #' @param x A vector or list
 #' @param quote_fun Quoting function, defaults to `backquote`. If set to `NULL`,
@@ -10,35 +13,17 @@
 #'
 #' @return A `character` vector
 #'
-#' @keywords quote
-#' @family quote
-#'
-#' @examples
-#' enumerate(c("one", "two", "three"))
-#'
-#' enumerate(c(1, 2, 3), quote_fun = NULL)
+#' @keywords deprecated
+#' @family deprecated
 #'
 #' @export
 enumerate <- function(x, quote_fun = backquote, conjunction = "and") {
-  lifecycle::deprecate_soft(
+  lifecycle::deprecate_stop(
     when = "1.1.0",
     what = "admiraldev::enumerate()",
     details = "This function was primarily used in error messaging, and can be
                replaced with 'cli' functionality: `cli::cli_abort('{.val {letters[1:3]}}')`"
   )
-
-  if (is.null(quote_fun)) {
-    quote_fun <- function(x) x
-  }
-  if (length(x) == 1L) {
-    quote_fun(x)
-  } else {
-    paste(
-      paste0(quote_fun(x[-length(x)]), collapse = ", "),
-      conjunction,
-      quote_fun(x[length(x)])
-    )
-  }
 }
 
 #' Wrap a String in Backquotes