+++ tests/testthat/test-dev_utilities.R
@@ -1,14 +1,10 @@
 # arg_name ----
-## Test 1: arg_name works ----
-test_that("arg_name Test 1: arg_name works", {
-  withr::local_options(lifecycle_verbosity = "quiet")
-  expect_equal(arg_name(sym("a")), "a")
-  expect_equal(arg_name(call("enquo", sym("a"))), "a")
-  expect_equal(arg_name(quote(enexpr(my_var))), "my_var")
-  expect_equal(arg_name(quote(rlang::enexpr(my_var))), "my_var")
-  expect_equal(arg_name(quote(mean(my_var))), "my_var")
-  expect_equal(arg_name(quote(mean(identity(my_var)))), "my_var")
-  expect_error(arg_name("a"), "Could not extract argument name from")
+## Test 1: deprecation error if function is called ----
+test_that("arg_name Test 1: deprecation error if function is called", {
+  expect_error(
+    arg_name(sym("a")),
+    class = "lifecycle_error_deprecated"
+  )
 })
 
 # convert_dtm_to_dtc ----
@@ -108,8 +104,8 @@ test_that("extract_vars Test 11: works with calls", {
 })
 
 # %or% ----
-## Test 12: works ----
-test_that("%or% Test 12: works", {
+## Test 12: %or% works ----
+test_that("extract_vars Test 12: %or% works", {
   input <- dplyr::tribble(
     ~USUBJID, ~AVAL,
     "P01",    2,
@@ -120,4 +116,34 @@ test_that("%or% Test 12: works", {
       dplyr::select(-AVAL) %>%
       dplyr::mutate(AVAL = sqrt("4") %or% 2)
   )
+})
+
+## Test 13: %or% deprecation message if function is called ----
+test_that("extract_vars Test 13: %or% deprecation message if function is called", {
+  expect_snapshot({
+    input <- dplyr::tribble(
+      ~USUBJID, ~AVAL,
+      "P01",    2,
+    )
+    result <- input %>%
+      dplyr::select(-AVAL) %>%
+      dplyr::mutate(AVAL = sqrt("4") %or% 2)
+  })
+})
+
+# valid_time_units ----
+## Test 14: works as intended ----
+test_that("valid_time_units Test 14: works as intended", {
+  expect_equal(
+    valid_time_units(),
+    c("years", "months", "days", "hours", "minutes", "seconds")
+  )
+})
+
+
+## Test 15: deprecation message if function is called ----
+test_that("valid_time_units Test 15: deprecation message if function is called", {
+  expect_snapshot({
+    result <- valid_time_units()
+  })
 })